Fixed mistakes in deleteServerOfTarget and removed some redundant functions
